checkpoint commit. moved holon_operations_api to reference layer. everything compiles except sweetest. nothing runs.
