+++ crates/holons_core/src/lib.rs
@@ -5,6 +5,6 @@ pub mod reference_layer;
 pub mod utils;
 
 // pub use core_shared_objects::*;
-pub use core_shared_objects::holon_operations_api::*;
+pub use reference_layer::holon_operations_api::*;
 pub use reference_layer::*;
 // pub use utils::*;
